refactor: unify the protocol_state_to_string function
