+++ crates/near-mpc-contract-interface/src/types/state.rs
@@ -255,6 +255,132 @@ pub enum ProtocolContractState {
     Resharing(ResharingContractState),
 }
 
+fn params_to_string(output: &mut String, parameters: &ThresholdParameters) {
+    output.push_str("    Participants:\n");
+    for (account_id, id, info) in &parameters.participants.participants {
+        output.push_str(&format!("      ID {}: {} ({})\n", id, account_id, info.url));
+    }
+    output.push_str(&format!(
+        "    Threshold: {}\n",
+        parameters.threshold.value()
+    ));
+}
+
+/// Formats the protocol state for human-readable display.
+///
+/// This does not calculate automatic timeouts for key generation or resharing
+/// attempts; an instance is reported as "active" whenever `instance` is
+/// populated, regardless of whether it would be considered timed out on-chain.
+pub fn protocol_state_to_string(contract_state: &ProtocolContractState) -> String {
+    let mut output = String::new();
+    match contract_state {
+        ProtocolContractState::NotInitialized => {
+            output.push_str("Contract is not initialized\n");
+        }
+        ProtocolContractState::Initializing(state) => {
+            output.push_str("Contract is in Initializing state (key generation)");
+            output.push_str(&format!("  Epoch: {}\n", state.generating_key.epoch_id));
+            output.push_str("  Domains:\n");
+            for (i, domain) in state.domains.domains.iter().enumerate() {
+                output.push_str(&format!("    Domain {}: {:?}, ", domain.id, domain.curve));
+                #[expect(clippy::comparison_chain)]
+                if i < state.generated_keys.len() {
+                    output.push_str(&format!(
+                        "key generated (attempt ID {})\n",
+                        state.generated_keys[i].attempt
+                    ));
+                } else if i == state.generated_keys.len() {
+                    output.push_str("generating key: ");
+                    if let Some(instance) = state.generating_key.instance.as_ref() {
+                        output.push_str(&format!(
+                            "active; current attempt ID: {}\n",
+                            instance.attempt_id
+                        ));
+                    } else {
+                        output.push_str(&format!(
+                            "not active; next attempt ID: {}\n",
+                            state.generating_key.next_attempt_id
+                        ));
+                    }
+                } else {
+                    output.push_str("queued for generation\n");
+                }
+            }
+            output.push_str("  Parameters:\n");
+            params_to_string(&mut output, &state.generating_key.parameters);
+            output.push_str("  Warning: this tool does not calculate automatic timeouts for key generation attempts\n");
+        }
+        ProtocolContractState::Running(state) => {
+            output.push_str("Contract is in Running state\n");
+            output.push_str(&format!("  Epoch: {}\n", state.keyset.epoch_id));
+            output.push_str("  Keyset:\n");
+            for (domain, key) in state
+                .domains
+                .domains
+                .iter()
+                .zip(state.keyset.domains.iter())
+            {
+                output.push_str(&format!(
+                    "    Domain {}: {:?}, key from attempt {}\n",
+                    domain.id, domain.curve, key.attempt
+                ));
+            }
+            output.push_str("  Parameters:\n");
+            params_to_string(&mut output, &state.parameters);
+        }
+        ProtocolContractState::Resharing(state) => {
+            output.push_str("Contract is in Resharing state\n");
+            output.push_str(&format!(
+                "  Epoch transition: original {} --> prospective {}\n",
+                state.previous_running_state.keyset.epoch_id, state.resharing_key.epoch_id,
+            ));
+            output.push_str("  Domains:\n");
+            for (i, domain) in state
+                .previous_running_state
+                .domains
+                .domains
+                .iter()
+                .enumerate()
+            {
+                output.push_str(&format!(
+                    "    Domain {}: {:?}, original key from attempt {}, ",
+                    domain.id, domain.curve, state.previous_running_state.keyset.domains[i].attempt
+                ));
+
+                #[expect(clippy::comparison_chain)]
+                if i < state.reshared_keys.len() {
+                    output.push_str(&format!(
+                        "reshared (attempt ID {})\n",
+                        state.reshared_keys[i].attempt
+                    ));
+                } else if i == state.reshared_keys.len() {
+                    output.push_str("resharing key: ");
+                    if let Some(instance) = state.resharing_key.instance.as_ref() {
+                        output.push_str(&format!(
+                            "active; current attempt ID: {}\n",
+                            instance.attempt_id
+                        ));
+                    } else {
+                        output.push_str(&format!(
+                            "not active; next attempt ID: {}\n",
+                            state.resharing_key.next_attempt_id
+                        ));
+                    }
+                } else {
+                    output.push_str("queued for resharing\n");
+                }
+            }
+            output.push_str("  Previous Parameters:\n");
+            params_to_string(&mut output, &state.previous_running_state.parameters);
+            output.push_str("  Proposed Parameters:\n");
+            params_to_string(&mut output, &state.resharing_key.parameters);
+
+            output.push_str("  Warning: this tool does not calculate automatic timeouts for resharing attempts\n");
+        }
+    }
+    output
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -264,5 +390,18 @@ mod tests {
         let state = ProtocolContractState::NotInitialized;
         let json = serde_json::to_string(&state).unwrap();
         assert!(json.contains("NotInitialized"));
+    }
+
+    #[test]
+    #[expect(non_snake_case)]
+    fn protocol_state_to_string__should_describe_not_initialized_state() {
+        // Given
+        let state = ProtocolContractState::NotInitialized;
+
+        // When
+        let output = protocol_state_to_string(&state);
+
+        // Then
+        assert_eq!(output, "Contract is not initialized\n");
     }
 }
